+++ src/charset/to_utf32.h
@@ -9,6 +9,6 @@
 namespace charset
 {
 
-std::u32string toUTF32(Charset srcCharset, const void *data, i32 nbytes);
+std::u32string toUTF32(Charset srcCharset, const void *data, i64 nbytes);
 
 }
