+++ src/core/membuff.cpp
@@ -3,28 +3,28 @@
 #include <algorithm>
 
 
-void MemBuff::resize(i32 len) {
+void MemBuff::resize(i64 len) {
     ensureCapacity(len);
     size_ = len;
 }
 
-i32 MemBuff::read(i32 offset, void *to, i32 len) const {
+i64 MemBuff::read(i64 offset, void *to, i64 len) const {
     if (size_ <= offset) {
         return 0;
     }
 
-    const i32 readable = std::min(size_ - offset, len);
+    const i64 readable = std::min(size_ - offset, len);
     std::memcpy(to, beg_ + offset, readable);
     return readable;
 }
 
-i32 MemBuff::remove(i32 offset, i32 len) {
+i64 MemBuff::remove(i64 offset, i64 len) {
     if (size_ <= offset) {
         return 0;
     }
 
-    const i32 removable = std::min(size_ - offset, len);
-    const i32 endOffset = offset + removable;
+    const i64 removable = std::min(size_ - offset, len);
+    const i64 endOffset = offset + removable;
     if (endOffset < size_) {
         std::memmove(beg_ + offset, beg_ + endOffset, size_ - endOffset);
     }
@@ -32,13 +32,13 @@ i32 MemBuff::remove(i32 offset, i32 len) {
     return removable;
 }
 
-i32 MemBuff::take(i32 offset, void *to, i32 len) {
+i64 MemBuff::take(i64 offset, void *to, i64 len) {
     if (size_ <= offset) {
         return 0;
     }
 
-    const i32 takable = std::min(size_ - offset, len);
-    const i32 endOffset = offset + takable;
+    const i64 takable = std::min(size_ - offset, len);
+    const i64 endOffset = offset + takable;
     std::memcpy(to, beg_ + offset, takable);
     if (endOffset < size_) {
         std::memmove(beg_ + offset, beg_ + endOffset, size_ - endOffset);
@@ -47,7 +47,7 @@ i32 MemBuff::take(i32 offset, void *to, i32 len) {
     return takable;
 }
 
-void MemBuff::write(i32 offset, const void *data, i32 len) {
+void MemBuff::write(i64 offset, const void *data, i64 len) {
     ensureCapacity(offset + len);
 
     // 若当前字节数小于指定偏移,先用0字节把这段空白初始化
@@ -62,23 +62,23 @@ void MemBuff::write(i32 offset, const void *data, i32 len) {
     size_ = std::max(size_, offset + len);
 }
 
-void MemBuff::append(const void *from, i32 len) {
-    const i32 newbytecount = size_ + len;
+void MemBuff::append(const void *from, i64 len) {
+    const i64 newbytecount = size_ + len;
     ensureCapacity(newbytecount);
     std::memcpy(beg_ + size_, from, len);
     size_ = newbytecount;
 }
 
-i32 MemBuff::calcRealNewCapacity(i32 curCapacity, i32 newCapacity) {
-	constexpr i32 kInit = 256;
-    i32 n = curCapacity == 0 ? kInit : curCapacity;
+i64 MemBuff::calcRealNewCapacity(i64 curCapacity, i64 newCapacity) {
+	constexpr i64 kInit = 256;
+    i64 n = curCapacity == 0 ? kInit : curCapacity;
 	while (n < newCapacity) {
 		n *= 2;
 	}
 	return n;
 }
 
-void MemBuff::realloc(i32 newcapacity) {
+void MemBuff::realloc(i64 newcapacity) {
     capacity_ = newcapacity;
     byte *ptr = new byte[capacity_];
     std::memcpy(ptr, beg_, size_);
